% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCDBRecode.R
\name{NCDBRecode}
\alias{NCDBRecode}
\title{National Cancer Database Recode Project}
\usage{
NCDBRecode(df)
}
\description{
This package serves to recode the standard NCDB excel or stata .sav file for ease of survival analysis
}
